feat: modificacion del nombre del servlet MostrarEjerciciosDeCategorias (estaba en minuscula)
